refactor(agent): extract unified tool executor factory
